Push the /ingests API down as well
