+++ archive/archive_api/src/apis/__init__.py
@@ -1,5 +1,6 @@
 # -*- encoding: utf-8
 
 from .bags import api as bags_api
+from .ingests import api as ingests_api
 
-__all__ = ["bags_api"]
+__all__ = ["bags_api", "ingests_api"]
