+++ archive/archive_api/src/archive_api.py
@@ -6,7 +6,7 @@ import daiquiri
 from flask import Flask
 from flask_restplus import Api
 
-from apis import bags_api
+from apis import bags_api, ingests_api
 import config
 from responses import ContextResponse
 from progress_manager import ProgressManager
@@ -38,6 +38,7 @@ app.config["PROGRESS_MANAGER"] = ProgressManager(
 )
 
 api.add_namespace(bags_api)
+api.add_namespace(ingests_api)
 
 # We can't move this import to the top because the views need the ``api``
 # instance defined in this file.
